poca cosa

+++ final/src/main/java/com/example/demo/repository/Repository.java
@@ -15,7 +15,7 @@ public interface Repository extends CrudRepository <Turbina, Long> {
     //List<Turbina> getAllInfoQuery();
 
     @Query("INSERT INTO TURBINA VALUES(:ID, :ANGULO, :ALTURA, :VELOCIDAD_MAX, :ENCENDIDO, :CARGA")
-    int addTurbina(@Param("ID")int id, @Param("ANGULO")int angulo, @Param("ALTURA")int altura, @Param("VELOCIDAD_MAX")int velocidad_max, @Param("ENCENDIDO")boolean on);
+    int addTurbina(@Param("ID")Long id, @Param("ANGULO")Long angulo, @Param("ALTURA")Long altura, @Param("VELOCIDAD_MAX")Long velocidad_max, @Param("ENCENDIDO")boolean on);
 
     // @Query("UPDATE TURBINA SET ANGULO = :ANGULO, ALTURA = :ALTURA, VELOCIDAD_MAX = :VELOCIDAD_MAX, ENCENDIDO = :ENCENDIDO WHERE ID = :ID")
     // int putInfo(@Param("ID")int i, @Param("ANGULO")int angulo, @Param("ALTURA")int altura, @Param("VELOCIDAD_MAX")int velocidad_max, @Param("ENCENDIDO")boolean on);
